refactor(web): stop exporting everything from network
Related to https://github.com/openSUSE/agama/issues/1049
